requirements.txt
### Proxcard MQTT Authentication Requirements
pyyaml
paho-mqtt
evdev
